Consistent reference from_raw, cast_local and null methods
This does a pass over all reference types to ensure they have consistent
`from_raw`, `cast_local` and `null` methods.

The `from_raw` methods now all require an `env: Env<'local>` argument so
we can guarantee that the returned wrapper is tied to a valid local
reference frame. This ensures the API is only used with valid local
references and can't be used to create wrappers that have a `'static`
lifetime. (`Env::global_from_raw` can instead be used for global
references that need a `'static` lifetime).

Requiring an `env` argument for `from_raw` methods is also important for
being able to assume that if code ever observes a `Reference` type with
a `'local` lifetime then that should be enough to prove that 1)
JavaVM::singleton() is initialized and 2) the current thread must be
attached to the JVM.

Even though there is a `Reference::null()` method, we add an explicit
`null()` method for each reference type that is `const`.

The descriptions for all reference types are now also consistent and
refer to the `java.*` type that they wrap.