Add ScopeTokens + consistent JavaVM/Env::with[_top]_local_frame APIs
This does a pass over the JavaVM `AttachGuard` APIs to mitigate risks with
acquiring `'static` guards.

The unsafe `JavaVM::get_env_attachment` and
`JavaVM::attach_current_thread_guard` APIs have been updated to take a
`&mut ScopeToken` reference which is where the lifetime for the
`AttachGuard` comes from.

A `ScopeToken` is a non-Send/non-Sync type that has no `const`
constructor and it borrowed with a mutable reference. All combined that
makes it _much_ less likely that someone could unwittingly acquire a
`'static` `ScopeToken` and therefore acquire a `'static` `AttachGuard`
that could be moved off the stack.

Although the `ScopeToken` type is only relevant for the `unsafe`
`AttachGuard` APIs, it helps make these APIs easier to use correctly
when considering that it's highly likely that applications will have a
static `JavaVM` and the `AttachGuard` lifetime previously came from
borrowing the `JavaVM`.

While reviewing the safe JavaVM APIs for running code with an Env
reference (and a hidden AttachGuard), it made sense to create a symmetry
with the Env API, considering:

- JavaVM::with_env_with_capacity was equivalent to Env::with_local_frame but
  the naming could make that clearer.

- JavaVM::with_current_env_frame is getting used in places that already
  have an `env` reference so it's redundant for them to use
  `env.get_java_vm()` and then use a JavaVM API which will then
  internally have to check for a valid Env pointer that we already had
  in the first place.

It's also noted that the need for a `&mut self` when calling
`Env::with_local_frame()` is a redundant restriction since this API does
not return a local reference and can also be (safely) bypassed by
calling the equivalent JavaVM API.

One arguable benefit from requiring a `&mut self` argument is that it blocks
some patterns of API misuse where someone could materialize multiple
`&mut Env` references for themselves. Considering that this is
already possible via the JavaVM thread attachment APIs and that we now
have fail-safe runtime checks (that will panic to avoid unsound behaviour)
then it should be reasonable to remove this restriction.

This renames `JavaVM::with_env_with_capacity` and
`JavaVM::with_current_env_frame` so we have symmetry with the Env API
like:

- JavaVM::with_local_frame
- JavaVM::with_top_local_frame
- Env::with_top_local_frame
- Env::with_local_frame

This removes `JavaVM::with_env` which was effectively just wrapper over
`JavaVM::with_local_frame` with a default capacity.

The name `with_top_local_frame` seemed clearer and more consistent with
`Env::with_local_frame` than `with_current_env_frame`. The documentation
refers to the "top" JNI stack frame and that seems more explicit than
"current".

Env::with_local_frame now works with a shared `&self` reference.

Code that was using `JavaVM::with_current_env_frame` while it already
had access to an `Env` reference has been updated to use
`Env::with_top_local_frame()` - avoiding a redundant thread attachment
check.

This makes a pass over the JavaVM and Env documentation to try and
clearly explain the design and provide guidance on how to avoid any risk
of being exposed to the runtime checks for the top JNI stack frame
(that can panic).

Fixes: #665