Mark `MonitorGuard` with `#[must_use]`
Because unlike `MutexGuard` from std there is no need to refer to this
guard, this makes it very easy to forget to store it in a variable
and thus unlock the object sooner than intended. Marking the type with
`#[must_use]` makes rustc generate a warning in such a case.