make env::NativeMethod transparent JNINativeMethod wrapper
This makes env::NativeMethod a transparent wrapper for JNINativeMethod
with an `unsafe` `::from_raw_parts()` constructor.

This makes it clearer what the safety concerns are with constructing a
`NativeMethod` descriptor and also avoids the need to allocate an
intermediate Vec within `Env::register_native_methods()`

There is also a plan to implement a `native_method!()` macro that can
safely construct a `NativeMethod` by guaranteeing that the function
matches the given JNI signature.

Even with `::from_raw_parts()` being `unsafe` we still _also_ keep
`register_native_methods()` as `unsafe` due to the risk of muddling up
the signature for static vs instance methods, which won't be something
that can verified at compile time just via a `native_method!` macro.