import jni_fn as jni_mangle macro
This imports the jni_fn macro from https://gitlab.com/antonok/jni_fn/
renamed to jni_mangle.

The plan is to use this internally within a `bind_java_type!{}` macro
but we need to make some tweaks to the interface.

The plan is to accept the function name and signature as optional second
and third arguments like:

`#[jni_mangle("foo.Bar", "nativeMethod", "(I)V")`

This ensures the macro can handle any valid Java method name (not just names
that are valid as Rust identifiers).

Accepting the signature also ensures it's able to mangle overloaded
methods.

The plan is to also use `export_name` instead of `no_mangle` so that the
non-mangled function name remains usable from the Rust crate.

Co-authored-by: Anton Lazarev <antonok35@gmail.com>