Make JNIStr::from_cstr fallible + remove 'From<CStr> for JNIStr'
JNIStr::from_cstr now returns an Option instead of panicking.

(It returns an `Option` instead of a `Result` so that it can still be a
const function that can run at compile time)

Since we now have the `jni_str!` macro this also removes the `From<CStr>
for JNIStr` implementation and migrates away from passing `CStr` literals
to JNI in favor of using `jni_str!`.

`jni_str!` is a better solution than casting `CStr` literals because it
has full unicode support and guarantees that the input string is
validated and encoded to MUTF-8 at compile time.