jni-macros: add `jni_sig!` + `jni_sig_[c,j]str!` macros
# `jni_sig` macro

This macro enables compile-time parsing of a method or field signature
and outputs a `MethodSignature` or `FieldSignature` struct that includes
a literal signature string (encoded as MUTF-8) and maps method
arguments, return values and field types into `JavaType` enum variants.

The macro can either be passed a standard JNI signature as input (e.g.
"(ILjava/lang/String;)Z") or it accepts a syntax like this:

- `(jint, JString) -> jboolean`
- `(arg0: JObject, arg1: bool) -> JObject`
- `(MyType, byte[][]) -> int[]`
- `(com.example.JavaType, bool) -> com.example.OtherType`

# `jni_sig_str`, `jni_sig_cstr` + `jni_sig_jstr` macros

Accept the same input as `jni_sig` but instead of outputting a struct
they just output a string literal for the signature, either as a plain
`&'static str`, a `&'static CStr` or a `&'static JNIStr` respectively.

# Groundwork for `bind_java_type` + `native_method` macros

The plan is to build on this same signature parsing syntax in follow up
work to implement a `bind_java_type!` and `native_method!` macros.

Although it's overkill for this initial `jni_sig!` macro, this macro
includes type mapping functionality that will be shared with the
`bind_java_type` and `native_method` macros in order to support
signatures that refer to rust types.