jni-macros: add native_method macros
This makes it possible to construct an `env::NativeMethod` with a
guarantee that the function pointer matches the JNI signature.

The macro has a long-form `prop = <val>,` config syntax and a short-form
signature syntax for common cases.

The signature parsing is based on the syntax introduced for `jni_sig!`

For example:

```
let method = native_method! {
    static extern fn MyRustType::my_native_method(arg: jint) -> bool
};
```

In this case `my_native_method` is mapped to "myNativeMethod" as the
Java method name. The signature would be "(I)Z" and the implementation
function would be `MyRustType::my_native_method`. If it was an instance
method then `MyRustType` would also be used for the `this` argument.