jni-macros: add native_method macro
This makes it possible to safely construct an `env::NativeMethod` with a
guarantee that the function pointer matches the JNI signature and a
(one-time) runtime check that the method receiver type is correct (i.e.
that it's an instance or static method).

The macro can optionally export a JNI mangled symbol like
"Java_com_example_MyType_myMethod" - which makes this a safer
alternative to the `#[jni_mangle]` attribute macro.

By default the macro will unwrap a Result with an ErrorPolicy as well as catch
panics which also makes it more ergonomic than `#[jni_mangle]`

The macro has a long-form `prop = <val>,` config syntax and a short-form
signature syntax for common cases.

The signature parsing is based on the syntax introduced for `jni_sig!`

For example:

```
let method = native_method! {
    java_type = com.example.MyType,
    static extern fn MyRustType::my_native_method(arg: jint) -> bool
};
```

In this case `my_native_method` is mapped to "myNativeMethod" as the
Java method name. The signature would be "(I)Z" and the implementation
function would be `MyRustType::my_native_method`. If it was an instance
method then `MyRustType` would also be used for the `this` argument.
Since `extern` says we want an exported symbol then a `java_type`
needs to be provided in order to generate the mangled name.