refs: Add `TransparentReference` marker trait
The Reference trait requires that the Kind and GlobalKind associated
types are always transparent JNI reference wrappers but the Reference
type itself isn't always required to be a transparent wrapper.

There are times though when it's useful to be able to assert that we have a
Reference type that is itself a transparent JNI reference wrapper and so
this adds a marker trait that is automatically implemented for all
Reference types where `Self == Self::Kind`.

For example this can be useful to assert when implementing an `AsRef`
cast into some Reference type in terms of an `unsafe` pointer cast.