Require pre-parsed JNI signatures
JNI signatures are now expected to be pre-parsed into a MethodSignature
or FieldSignature before being used to make method calls or get/set
fields.

In the common case the `jni_sig!()` macro can be used to parse a JNI
signature at compile time and emit the appropriate `*Signature` struct.

In the unlikely case that a signature needs to be parsed at runtime then
`RuntimeMethodSignature/RuntimeFieldSignature::from_str` can be used
and it's then possible to borrow a `MethodSignature` or `FieldSignature`
from these runtime structs.